separate password form field to a shared widget
