+++ lib/pages/passwordUpdate/oldPasswordConfirmForm.dart
@@ -5,6 +5,7 @@ import 'package:readr_app/blocs/passwordUpdate/bloc.dart';
 import 'package:readr_app/blocs/passwordUpdate/events.dart';
 import 'package:readr_app/blocs/passwordUpdate/states.dart';
 import 'package:readr_app/helpers/dataConstants.dart';
+import 'package:readr_app/pages/shared/passwordFormField.dart';
 import 'package:readr_app/pages/shared/passwordValidatorWidget.dart';
 
 class OldPasswordConfirmForm extends StatefulWidget {
@@ -18,7 +19,6 @@ class OldPasswordConfirmForm extends StatefulWidget {
 
 class _OldPasswordConfirmFormState extends State<OldPasswordConfirmForm> {
   final _passwordEditingController = TextEditingController();
-  bool _isHidden = true;
   bool _passwordIsValid = false;
 
   @override
@@ -41,12 +41,6 @@ class _OldPasswordConfirmFormState extends State<OldPasswordConfirmForm> {
 
   bool _isPasswordValid() {
     return _passwordEditingController.text != null && _passwordEditingController.text.length >= 6;
-  }
-
-  void _togglePasswordView() {
-    setState(() {
-      _isHidden = !_isHidden;
-    });
   }
 
   @override
@@ -74,7 +68,10 @@ class _OldPasswordConfirmFormState extends State<OldPasswordConfirmForm> {
         SizedBox(height: 24),
         Padding(
           padding: const EdgeInsets.only(left: 24.0, right: 24.0),
-          child: _passwordField(),
+          child: PasswordFormField(
+            title: '目前的密碼',
+            passwordEditingController: _passwordEditingController,
+          ),
         ),
         SizedBox(height: 8),
         Padding(
@@ -103,32 +100,6 @@ class _OldPasswordConfirmFormState extends State<OldPasswordConfirmForm> {
         ),
         SizedBox(height: 24),
       ],
-    );
-  }
-
-  Widget _passwordField() {
-    return Form(
-      child: TextFormField(
-        controller: _passwordEditingController,
-        obscureText: _isHidden,
-        style: TextStyle(
-          color: Colors.black,
-          fontSize: 17,
-        ),
-        decoration: InputDecoration(
-          suffixIcon: IconButton(
-            onPressed: () => _togglePasswordView(),
-            icon: _isHidden 
-            ? Icon(Icons.visibility)
-            : Icon(Icons.visibility_off),
-          ),
-          labelText: '目前的密碼',
-          labelStyle: TextStyle(
-            color: Colors.black,
-            fontSize: 17,
-          ),
-        ),
-      ),
     );
   }
